[PRW 2.0] Moved rw2 proto to the full path (both package name and placement)
This is what buf recommends https://buf.build/docs/reference/protobuf-files-and-packages#packages
