+++ prompb/io/prometheus/write/v2/types.proto
@@ -1,4 +1,4 @@
-// Copyright 2017 Prometheus Team
+// Copyright 2024 Prometheus Team
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,58 +11,55 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// NOTE: This file is also available on https://buf.build/prometheus/prometheus/docs/main:io.prometheus.write.v2
+
 syntax = "proto3";
-package write.v2;
+package io.prometheus.write.v2;
 
 option go_package = "writev2";
 
 import "gogoproto/gogo.proto";
 
-// WriteRequest represents a Remote Write 2.0 request to write given time series
-// to remote destination. Note that Remote Write 2.0 requires a content
-// negotiation for version and compressions, explained in
-// https://docs.google.com/document/d/1jx1fqpRnM0pAndeo3AgY7g6BLxN3Ah8R0Mm8RvNsHoU/edit
-// TODO(bwplotka): Change URL to Prometheus docs once ready.
-message WriteRequest {
-  // symbols contains de-duplicated array of string elements used for various
-  // items in WriteRequest like labels and some metadata items. To decode
+// Request represents a request to write the given timeseries to a remote destination.
+// This was introduced in the Remote Write 2.0 specification:
+// https://prometheus.io/docs/concepts/remote_write_spec_2_0/
+// The canonical Content-Type request header value for this message is
+// "application/x-protobuf;proto=io.prometheus.write.v2.Request"
+//
+// NOTE: gogoproto options might change in future for this file, they
+// are not part of the spec proto (they only modify the generated Go code, not
+// the serialized message). See: https://github.com/prometheus/prometheus/issues/11908
+message Request {
+  // symbols contains a de-duplicated array of string elements used for various
+  // items in a Request message, like labels and metadata items. To decode
   // each of those items, referenced, by "ref(s)" suffix, you need to lookup the
   // actual string by index from symbols array. The order of strings is up to
   // the client, server should not assume any particular encoding.
   repeated string symbols = 1;
-  // timeseries represents array of distinct series with 0 or more samples.
+  // timeseries represents an array of distinct series with 0 or more samples.
   repeated TimeSeries timeseries = 2 [(gogoproto.nullable) = false];
 }
 
 // TimeSeries represents a single series.
 message TimeSeries {
   // labels_refs is a list of label name-value pair references, encoded
-  // as indices to the WriteRequest.symbols array. This list's len is always
-  // a multiple of 2, and the underlying labels should be sorted.
+  // as indices to the Request.symbols array. This list's length is always
+  // a multiple of two, and the underlying labels should be sorted.
   //
   // Note that there might be multiple TimeSeries objects in the same
-  // WriteRequests with the same labels e.g. for different exemplars, metadata
+  // Requests with the same labels e.g. for different exemplars, metadata
   // or created timestamp.
   repeated uint32 labels_refs = 1;
 
-  // samples contain zero or more samples for a given timeseries. For typical
-  // clients, in healthy cases, there will be only one sample, for ~real
-  // time metric streaming. Samples can, in theory, co-exist with histogram samples
-  // (histograms field), although it should be extremely rare in practice (e.g.
-  // only when classic histogram series and native histogram share exactly the
-  // same metric name).
+  // Timeseries messages can either specify samples or (native) histogram samples
+  // (histogram field), but not both. For typical clients (~real-time metric
+  // streaming), in healthy cases, there will be only one sample or histogram.
   //
-  // Samples are sorted by timestamp (older first).
+  // Samples and histograms are sorted by timestamp (older first).
   repeated Sample samples = 2 [(gogoproto.nullable) = false];
-  // histograms contain zero or more histogram samples for a given timeseries.
-  // For typical clients, in healthy cases, there will be only one sample, for ~real
-  // time metric streaming. histograms can co-exist with samples (see samples
-  // for details).
-  //
-  // histograms are sorted by timestamp (older first).
   repeated Histogram histograms = 3 [(gogoproto.nullable) = false];
 
-  // exemplars represents optional set of exemplars attached to this series' samples.
+  // exemplars represents an optional set of exemplars attached to this series' samples.
   repeated Exemplar exemplars = 4 [(gogoproto.nullable) = false];
 
   // metadata represents the metadata associated with the given series' samples.
@@ -71,15 +68,15 @@ message TimeSeries {
   // created_timestamp represents an optional created timestamp associated with
   // this series' samples in ms format, typically for counter or histogram type
   // metrics. Note that some servers might require this and in return fail to
-  // ingest such series within the WriteRequest.
+  // ingest such samples within the Request.
   //
   // For Go, see github.com/prometheus/prometheus/model/timestamp/timestamp.go
   // for conversion from/to time.Time to Prometheus timestamp.
   //
-  // NOTE: Optional key word is omitted due to
+  // Note that the "optional" keyword is omitted due to
   // https://cloud.google.com/apis/design/design_patterns.md#optional_primitive_fields
   // Zero value means value not set. If you need to use exactly zero value for
-  // timestamp, use 1 millisecond before or after.
+  // the timestamp, use 1 millisecond before or after.
   int64 created_timestamp = 6;
 }
 
@@ -88,7 +85,7 @@ message TimeSeries {
 // the metric changes.
 message Exemplar {
   // labels_refs is a list of label name-value pair references, encoded
-  // as indices to the WriteRequest.symbols array. This list's len is always
+  // as indices to the Request.symbols array. This list's len is always
   // a multiple of 2, and the underlying labels should be sorted.
   repeated uint32 labels_refs = 1;
   // value represents an exact example value. This can be useful when the exemplar
@@ -97,11 +94,11 @@ message Exemplar {
   // timestamp represents an optional timestamp of the example in ms.
   // For Go, see github.com/prometheus/prometheus/model/timestamp/timestamp.go
   // for conversion from/to time.Time to Prometheus timestamp.
-  // 
-  // NOTE: Optional key word is omitted due to
+  //
+  // Note that the "optional" keyword is omitted due to
   // https://cloud.google.com/apis/design/design_patterns.md#optional_primitive_fields
   // Zero value means value not set. If you need to use exactly zero value for
-  // timestamp, use 1 millisecond before or after.
+  // the timestamp, use 1 millisecond before or after.
   int64 timestamp = 3;
 }
 
@@ -128,10 +125,10 @@ message Metadata {
     METRIC_TYPE_STATESET       = 7;
   }
   MetricType type = 1;
-  // help_ref is a reference to the WriteRequest.symbols array representing help
+  // help_ref is a reference to the Request.symbols array representing help
   // text for the metric.
   uint32 help_ref = 3;
-  // unit_ref is a reference to the WriteRequest.symbols array representing unit
+  // unit_ref is a reference to the Request.symbols array representing unit
   // for the metric.
   uint32 unit_ref = 4;
 }
