+++ scripts/genproto.sh
@@ -42,7 +42,7 @@ for dir in ${DIRS}; do
             ./*.proto
 		protoc --gogofast_out=plugins=grpc:. -I=. \
             -I="${GOGOPROTO_PATH}" \
-            ./write/v2/*.proto
+            ./io/prometheus/write/v2/*.proto
 		protoc --gogofast_out=Mgoogle/protobuf/timestamp.proto=github.com/gogo/protobuf/types,paths=source_relative:. -I=. \
             -I="${GOGOPROTO_PATH}" \
             ./io/prometheus/client/*.proto
