+++ storage/remote/write_handler.go
@@ -24,6 +24,8 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 
+	writev2 "github.com/prometheus/prometheus/prompb/io/prometheus/write/v2"
+
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
 
@@ -33,7 +35,6 @@ import (
 	"github.com/prometheus/prometheus/model/exemplar"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/prompb"
-	writev2 "github.com/prometheus/prometheus/prompb/write/v2"
 	"github.com/prometheus/prometheus/storage"
 	otlptranslator "github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite"
 )
@@ -204,7 +205,7 @@ func (h *writeHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		err = h.write(r.Context(), &req)
 	case RemoteWriteVersion20HeaderValue:
 		// 2.0 request.
-		var reqMinStr writev2.WriteRequest
+		var reqMinStr writev2.Request
 		if err := proto.Unmarshal(decompressed, &reqMinStr); err != nil {
 			level.Error(h.logger).Log("msg", "Error decoding remote write request", "err", err.Error())
 			http.Error(w, err.Error(), http.StatusBadRequest)
@@ -452,7 +453,7 @@ func (h *otlpWriteHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 }
 
-func (h *writeHandler) writeMinStr(ctx context.Context, req *writev2.WriteRequest) (err error) {
+func (h *writeHandler) writeMinStr(ctx context.Context, req *writev2.Request) (err error) {
 	outOfOrderExemplarErrs := 0
 
 	app := h.appendable.Appender(ctx)
